Consider rubocop installed if bundled
For those that vendorize their gems, the rubocop binary may not be on
the exec path, even if it has been bundled. For such users, their check
command will need to be `bundle exec rubocop` instead. This change will
allow rubocop-emacs to proceed with using the bundled rubocop if rubocop
is not found on the exec path.

Fixes https://github.com/bbatsov/rubocop-emacs/issues/17